merfe Binary and TypeAtom
